add img upload & storage, download invoice

+++ controllers/admin.js
@@ -1,6 +1,17 @@
 import { validationResult } from "express-validator";
+import fs from "fs";
+import { S3 } from "@aws-sdk/client-s3";
 import Product from "../models/product.js";
-import { extractFlashMessage } from "../utils/helpers.js";
+import {
+  deleteFile,
+  extractFlashMessage,
+  rootPath,
+  toArrayBuffer,
+} from "../utils/helpers.js";
+
+const s3 = new S3({
+  region: "ap-southeast-1",
+});
 
 export const getAddProduct = (req, res, next) => {
   const errorMessage = extractFlashMessage(req, "error");
@@ -16,6 +27,7 @@ export const getAddProduct = (req, res, next) => {
 
 export const postAddProduct = async (req, res, next) => {
   const errors = validationResult(req);
+  const imageFile = req.file;
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
@@ -26,14 +38,35 @@ export const postAddProduct = async (req, res, next) => {
       oldInput: req.oldInput,
     });
   }
+  if (!imageFile) {
+    return res.status(422).render("admin/edit-product", {
+      pageTitle: "Add Product",
+      path: "/admin/add-product",
+      editing: false,
+      errorMessage: "Attached file is not an image.",
+      validationErrors: [],
+      oldInput: req.oldInput,
+    });
+  }
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+
+  // upload image to s3
+  const slug = imageFile.filename;
+  const img = fs.readFileSync(imageFile.path);
+  const bufferedImage = toArrayBuffer(img);
+  s3.putObject({
+    Bucket: process.env.AWS_BUCKET_NAME,
+    Key: slug,
+    Body: Buffer.from(bufferedImage),
+    ContentType: imageFile.mimetype,
+  });
+
   const price = req.body.price;
   const description = req.body.description;
   const product = new Product({
     title,
     price,
-    imageUrl,
+    imageUrl: `${process.env.AWS_IMAGE_STORAGE_URL}/${slug}`,
     description,
     userId: req.user,
   });
@@ -67,6 +100,7 @@ export const getEditProduct = async (req, res, next) => {
 
 export const postEditProduct = async (req, res, next) => {
   const errors = validationResult(req);
+  const imageFile = req.file;
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
@@ -77,7 +111,7 @@ export const postEditProduct = async (req, res, next) => {
       oldInput: req.oldInput,
     });
   }
-  const { productId, title, imageUrl, price, description } = req.body;
+  const { productId, title, price, description } = req.body;
   const updatedProduct = await Product.findOne({
     _id: productId,
     userId: req.user._id,
@@ -85,9 +119,30 @@ export const postEditProduct = async (req, res, next) => {
   if (!updatedProduct) {
     return res.redirect("/admin/products");
   }
+  if (imageFile) {
+    // delete old image from s3 and tmp folder
+    const oldImage = updatedProduct.imageUrl.split(
+      `${process.env.AWS_IMAGE_STORAGE_URL}/`
+    )[1];
+    deleteFile(rootPath("tmp", "images", oldImage), next);
+    s3.deleteObject({
+      Bucket: process.env.AWS_BUCKET_NAME,
+      Key: oldImage,
+    });
+    // upload image to s3
+    const slug = imageFile.filename;
+    const img = fs.readFileSync(imageFile.path);
+    const bufferedImage = toArrayBuffer(img);
+    s3.putObject({
+      Bucket: process.env.AWS_BUCKET_NAME,
+      Key: slug,
+      Body: Buffer.from(bufferedImage),
+      ContentType: imageFile.mimetype,
+    });
+    updatedProduct.imageUrl = `${process.env.AWS_IMAGE_STORAGE_URL}/${slug}`;
+  }
   updatedProduct.title = title;
   updatedProduct.price = price;
-  updatedProduct.imageUrl = imageUrl;
   updatedProduct.description = description;
   await updatedProduct.save();
   res.redirect("/admin/products");
@@ -104,11 +159,24 @@ export const getProducts = async (req, res, next) => {
 
 export const postDeleteProduct = async (req, res, next) => {
   const { productId } = req.body;
+  const product = await Product.findById(productId);
+  if (!product) {
+    return res.redirect("/admin/products");
+  }
+  // delete image from s3 and tmp folder
+  const oldImage = product.imageUrl.split(
+    `${process.env.AWS_IMAGE_STORAGE_URL}/`
+  )[1];
+  deleteFile(rootPath("tmp", "images", oldImage), next);
+  s3.deleteObject({
+    Bucket: process.env.AWS_BUCKET_NAME,
+    Key: oldImage,
+  });
   await Product.deleteOne({ userId: req.user._id, _id: productId });
   res.redirect("/admin/products");
 };
 
-const adminController = {
+export default {
   getAddProduct,
   postAddProduct,
   getEditProduct,
@@ -116,5 +184,3 @@ const adminController = {
   getProducts,
   postDeleteProduct,
 };
-
-export default adminController;
